Move payroll rules' related models to its new namespace
