+++ app/Models/Settings/PayrollRules/IncomeTaxPayrollRules.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Settings\PayrollRules;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
